+++ app.py
@@ -1,0 +1,7 @@
+__author__ = "NR-SkaterBoy"
+__email__ = "nr.rick.dev@gmail.com"
+__website__ = "https://richardneuvald.tk/"
+__github__ = "https://github.com/NR-SkaterBoy"
+__issues__ = "https://github.com/NR-SkaterBoy/LinuxSystemUpdate/issues"
+__version__ = "0.8.0"
+__update__ = "2022.07.09"