+++ bash/lsu_update.sh
@@ -4,7 +4,9 @@
 # E-mail: nr.rick.dev@gmail.com
 # LSU update
 
+export SUDO_ASKPASS="bash/pass.sh"s
+
 git pull
 shell=True
-pip3 install -r requirements.txt
+sudo -A python3 install.py