+++ createfile.py
@@ -1,13 +1,16 @@
 import os
 import json
+# import shutil
 
 log_folder = f"/home/{os.getlogin()}/lsu/logs"
 files_folder = f"/home/{os.getlogin()}/lsu/files"
+# img_folder = f"/home/{os.getlogin()}/lsu/img/"
 
 modules_file = f"/home/{os.getlogin()}/lsu/files/modules.json"
 application_file = f"/home/{os.getlogin()}/lsu/files/app.json"
 
-os.mkdir(f"/home/{os.getlogin()}/lsu")
+if not os.path.isdir(f"/home/{os.getlogin()}/lsu"):
+    os.mkdir(f"/home/{os.getlogin()}/lsu")
 
 # LogFolder and neccesary file(s) [Modules]
 if (not os.path.isdir(log_folder)):
@@ -33,3 +36,16 @@ if (not os.path.isfile(application_file)):
     with open(application_file, "w") as app_file:
         app_file.write(json_object)
     json.dumps(app)
+
+# Image folder with images
+# if (not os.path.isdir(img_folder)):
+#     os.mkdir(os.path.join(img_folder))
+    
+
+# fetch all files
+# for file_name in os.listdir("img"):
+#     source = "img/" + file_name
+#     destination = img_folder + file_name
+#     if os.path.isfile(source):
+#         shutil.copy(source, destination)
+#         print('copied', file_name)